Change params immediately before upload
To facilitate file uploads that are part of a larger form, such as having notes describing the file, properties of the file, change log, etc., I needed to be able to pass parameters that might change well after the upload form was created.  Therefore, I moved the population of the hidden inputs from the main body of the constructor into the SubmitUpload() function.  The hidden inputs get overwritten with every new file that's added, for example to capture a textarea content that has changed on the page.  Example usage with a custom startupload callback...

startupload: function (SubmitUpload, e, data, settings)
{
	let form_data = form_data_function(page_id); // scrape a form on the page
	settings.params = form_data; // update the params to include the form contents
	if (!confirm('Uploading this file will submit all of your changes, are you sure?')) { data.abort(); return false; }
	SubmitUpload();
}