Added a sanitize filename function to the options
The optional sanitize function is called when a new file is added and whenever any filename input field loses focus.  It accepts the filename as a parameter and requires a filename to be returned.  This provides an opportunity to modify any filename according to your particular standards, such as stripping out special characters that will cause errors on your server's file system or database, forcing camel case, or appending a timestamp.